backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-dotenv
openai
psycopg2-binary